cache values casting

+++ src/adapters/tip_adapter.py
@@ -247,7 +247,8 @@ def get_cache_model(train_set, model, augment_epochs=10):
     train_images_features_agg /= train_images_features_agg.norm(dim=-1, keepdim=True)
     train_images_features_agg = train_images_features_agg.permute(1, 0)
 
-    cache_values = F.one_hot(torch.cat(cache_values, dim=0)).to(torch.float32)
+    cache_values = F.one_hot(torch.cat(cache_values, dim=0))
+    cache_values = cache_values.to(torch.float32)
     cache_keys = train_images_features_agg.to(torch.float32)
     assert cache_keys.shape[1] == cache_values.shape[0]
     assert cache_values.shape[1] == len(train_set.classes)
